requirements.txt
# ─── Embedding stack ─────────────────────────────────────────────
torch>=2.2.0                    # sentence‑transformers runs on Torch
sentence-transformers>=2.6.1    # MiniLM embeddings
faiss-cpu>=1.7.4                # vector index

# ─── API layer ───────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.8.0
python-multipart                # enables file‑upload endpoints

# ─── Ollama client + misc utilities ──────────────────────────────
requests>=2.31.0                # call the local Ollama server
tqdm>=4.66.0                    # progress bars during ingestion
pandas>=2.2.0                   # parse Excel knowledge base
openpyxl>=3.1.2                 # Excel engine for pandas
